test(sync): split post-merge-validator.test.ts by validator concern
Split the 3,216-LOC monolithic post-merge-validator test file into nine
concern-named files plus a shared validator-fixtures helper.

Per-concern split (7 named):
- tombstones (incl. ENTITY_FIELD_MAP, validateBucketContentTags)
- hierarchy (cycle detection, self-reference, orphan-parent)
- sort-order (tie partitioning, NULL_GROUP edge cases)
- check-in (record normalization + dispatch)
- friend-connection (status reconciliation, branch coverage)
- fronting-sessions (endTime/subject normalization)
- fronting-comments (author normalization)

Cross-cutting / branch coverage (2 additional):
- runall-cross-cutting: multi-validator interaction, onError plumbing
- runall-branch-coverage: timer/webhook config branch coverage

Helper extracts makeKeys, makeSessions, makeGroup, makeRegion,
makeFrontingSession, makeTimer, makeWebhookSession, makeActiveMember,
makeArchivedMember plus the sodium lifecycle.

All split files <=500 LOC. 951 sync tests green.
Original post-merge-validator.test.ts deleted.
